fix(arborist): workspaces respect overrides on subsequent installs (#8160)
<!-- What / Why -->

Fixes: https://github.com/npm/cli/issues/7660
https://github.com/npm/cli/issues/5443

<!-- Describe the request in detail. What it does and why it's being
changed. -->

Currently overrides are applied correctly to workspaces when a user does
their initial `npm install`. However, when a user runs `npm install`
again, the overrides are not being respected, and versions that the user
has specifically overridden because of vulnerabilities or other reasons,
are being installed in the node_modules of those workspaces. This
ensures that when a package-lock.json is loaded, the overrides are
calculated and applied to the workspaces.